split out getAttributes
